Align margins on tab edges, make margins user-configurable
Previously there was an extra space on the right side of the tabs that isn't present on the left. In addition to fixing the default alignment, this adds options for configuring the appearance of both sides.

By default, a space is added to each edge. If they're set to nil or an empty string then no space is added. Separate left/right options allow doing things like left="[" + right="]".

Also corrects the documentation for some of the settings.