Ensure that face-background reads indirect values
The old design would return an inappropriate nil value if the referenced
face did not have an explicit background attribute.  This can happen if
a theme uses ':inherit' instead of specifying the face attributes
directly.

By declaring a fallback value, we make sure that such indirection will
note return a nil value.

This is related to issue 178:
<https://github.com/ema2159/centaur-tabs/issues/178>.